feat: Add risk analysis and client data generator
This commit adds the functionality to generate risk analysis for a given client based on their CNPJ and the policies stored in the BigQuery. It also includes a client data generator function that generates fictitious client data based on a given CNPJ. This will be used to generate sample data for testing the risk analysis.